requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
yt-dlp==2024.12.13
python-multipart==0.0.12
pydantic==2.10.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
aiofiles==24.1.0
slowapi==0.1.9
requests==2.32.3
 